parameterization

+++ pytests/parameterization_camera.py
@@ -41,7 +41,7 @@ data = dataset.read(x=(0, 256), y=(0, 256), z=(0, 512))
 dtype = torch.float32
 data = data.astype(float)
 volume = torch.from_numpy(data).unsqueeze(0)
-volume = torch.tensor(volume, dtype=dtype, device=device)
+volume = torch.tensor(volume, dtype=torch.float16, device=device)
 X, Y, Z = 256, 256, 512
 
 torch.set_printoptions(sci_mode=False, precision=3)
